# core/vm: Simplify error handling in interpreter loop