--- a/core/vm/interpreter.go
+++ b/core/vm/interpreter.go
@@ -270,8 +270,6 @@ func (in *EVMInterpreter) Run(contract *Contract, input []byte, readOnly bool) (
 		}
 
 		switch {
-		case operation.reverts:
-			return res, ErrExecutionReverted
 		case operation.halts:
 			return res, nil
 		case !operation.jumps: