This simplifies interpreter loop exit condition to just `err != nil`. This catches:
- reverting instructions (i.e. just `REVERT`),
- halting instructions (by them returning `errStopToken`).

Also `operation.jumps` check is removed by setting `pc - 1` inside the jump instruction and doing `pc++` in the loop unconditionally.

Finally, instructions manipulating return data are handling this themselves. This eliminates another `operation.returns` check.