@chfast 2021-11-22
```
name                                  old time/op  new time/op  delta                                                                                                                                                             
micro/memory_grow_mload/nogrow         295µs ± 0%   280µs ± 1%   -5.08%  (p=0.000 n=9+10)                                                                                                                                         
micro/memory_grow_mload/by1            303µs ± 1%   288µs ± 1%   -4.80%  (p=0.000 n=10+9)                                                                                                                                         
micro/memory_grow_mload/by16           370µs ± 1%   355µs ± 0%   -4.14%  (p=0.000 n=10+10)                                                                                                                                        
micro/memory_grow_mload/by32           451µs ± 0%   436µs ± 0%   -3.22%  (p=0.000 n=9+9)                                                                                                                                          
micro/signextend/zero                  434µs ± 1%   420µs ± 0%   -3.12%  (p=0.000 n=10+9)                                                                                                                                         
micro/signextend/one                   443µs ± 1%   432µs ± 0%   -2.40%  (p=0.000 n=10+10)                                                                                                                                        
micro/jumpdests_0xffff/empty          3.44µs ± 1%  3.45µs ± 1%   +0.34%  (p=0.015 n=10+9)                                                                                                                                         
micro/loop_with_many_jumpdests/empty  3.59µs ± 1%  3.59µs ± 1%     ~     (p=0.713 n=8+10)                                                                                                                                         
micro/jump_around/empty                182µs ± 0%   175µs ± 0%   -4.04%  (p=0.000 n=9+9)                                                                                                                                          
micro/memory_grow_mstore/nogrow        530µs ± 0%   518µs ± 1%   -2.42%  (p=0.000 n=10+9)                                                                                                                                         
micro/memory_grow_mstore/by1           540µs ± 1%   528µs ± 0%   -2.34%  (p=0.000 n=9+8)                                                                                                                                          
micro/memory_grow_mstore/by16          624µs ± 0%   612µs ± 0%   -2.03%  (p=0.000 n=9+10)                                                                                                                                         
micro/memory_grow_mstore/by32          694µs ± 1%   681µs ± 1%   -1.91%  (p=0.000 n=9+10)                                                                                                                                         
micro/JUMPDEST_n0/empty               4.10ms ± 1%  3.52ms ± 0%  -14.28%  (p=0.000 n=10+9)                                                                                                                                         
main/sha1_divs/empty                   219µs ± 0%   208µs ± 0%   -4.92%  (p=0.000 n=8+10)                                                                                                                                         
main/sha1_divs/1351                   4.44ms ± 0%  4.23ms ± 1%   -4.70%  (p=0.000 n=10+10)
main/sha1_divs/2737                   8.64ms ± 0%  8.22ms ± 0%   -4.84%  (p=0.000 n=9+10)
main/sha1_divs/5311                   16.9ms ± 0%  16.1ms ± 0%   -5.02%  (p=0.000 n=10+9)
main/weierstrudel/0                    494µs ± 0%   481µs ± 0%   -2.66%  (p=0.000 n=9+10)
main/weierstrudel/1                   1.05ms ± 0%  1.01ms ± 0%   -3.35%  (p=0.000 n=10+10)
main/weierstrudel/3                   1.64ms ± 1%  1.59ms ± 0%   -3.49%  (p=0.000 n=10+9)
main/weierstrudel/9                   3.42ms ± 1%  3.31ms ± 0%   -3.28%  (p=0.000 n=10+10)
main/weierstrudel/14                  4.90ms ± 0%  4.74ms ± 0%   -3.22%  (p=0.000 n=9+10)
main/sha1_shifts/empty                 177µs ± 1%   170µs ± 0%   -3.69%  (p=0.000 n=10+6)
main/sha1_shifts/1351                 3.61ms ± 1%  3.50ms ± 0%   -3.11%  (p=0.000 n=10+10)
main/sha1_shifts/2737                 7.03ms ± 0%  6.82ms ± 1%   -2.97%  (p=0.000 n=9+10)
main/sha1_shifts/5311                 13.7ms ± 1%  13.3ms ± 0%   -3.08%  (p=0.000 n=10+10)
main/blake2b_huff/empty                111µs ± 1%   109µs ± 1%   -2.52%  (p=0.000 n=10+10)
main/blake2b_huff/2805nulls           2.14ms ± 0%  2.09ms ± 1%   -2.65%  (p=0.000 n=10+10)
main/blake2b_huff/5610nulls           4.18ms ± 1%  4.05ms ± 1%   -3.08%  (p=0.000 n=10+10)
main/blake2b_huff/8415nulls           6.09ms ± 0%  5.93ms ± 1%   -2.62%  (p=0.000 n=7+10)
main/blake2b_shifts/2805nulls         16.8ms ± 2%  15.5ms ± 1%   -7.88%  (p=0.000 n=10+10)
main/blake2b_shifts/5610nulls         33.3ms ± 1%  30.7ms ± 0%   -7.77%  (p=0.000 n=9+9)
main/blake2b_shifts/8415nulls         49.7ms ± 1%  45.8ms ± 1%   -7.91%  (p=0.000 n=9+10)
[Geo mean]                            1.13ms       1.09ms        -3.92%
```