@s1na 2021-11-23
So I built this branch and synced goerli from block 5,416,168 to the current head (5900905) without seeing an issue