@holiman 2021-11-24
Gave this a test on the existing benchmarks. I had expected it to have a bit of impact on `SimpleLoop/loop-100M-6` -- which is mainly a _lot_ trivial operations in a loop, burning through 100M gas. The effects are there, if you squint, but benchstat doesn't think it's statistically significant enough to report (I ran with `--count 5`). 
```
$ benchstat master.txt ipsilon.txt 
name                                   old time/op    new time/op    delta
SimpleLoop/staticcall-identity-100M-6     276ms ± 0%     277ms ± 1%    ~     (p=0.413 n=4+5)
SimpleLoop/call-identity-100M-6           358ms ± 5%     360ms ± 9%    ~     (p=0.548 n=5+5)
SimpleLoop/loop-100M-6                    521ms ± 9%     505ms ± 5%    ~     (p=0.310 n=5+5)
SimpleLoop/call-nonexist-100M-6           902ms ± 2%     908ms ± 3%    ~     (p=1.000 n=5+5)
SimpleLoop/call-EOA-100M-6                395ms ± 4%     394ms ± 5%    ~     (p=0.841 n=5+5)
SimpleLoop/call-reverting-100M-6          888ms ± 2%     879ms ± 3%    ~     (p=0.310 n=5+5)
```