@chfast 2021-11-24
### evmone benchmark suite on go 1.17
```
name                             old time/op  new time/op  delta                                                                                                                                                                  
micro/memory_grow_mload/nogrow    249µs ± 1%   239µs ± 0%   -4.10%  (p=0.000 n=10+9)                                                                                                                                              
micro/memory_grow_mload/by1       257µs ± 1%   247µs ± 1%   -4.18%  (p=0.000 n=10+10)                                                                                                                                             
micro/memory_grow_mload/by16      321µs ± 1%   311µs ± 1%   -3.29%  (p=0.000 n=10+10)                                                                                                                                             
micro/memory_grow_mload/by32      398µs ± 0%   390µs ± 1%   -2.07%  (p=0.000 n=10+10)                                                                                                                                             
micro/signextend/zero             463µs ± 0%   447µs ± 0%   -3.39%  (p=0.000 n=10+10)                                                                                                                                             
micro/signextend/one              471µs ± 1%   457µs ± 1%   -3.11%  (p=0.000 n=10+10)                                                                                                                                             
micro/jump_around/empty           154µs ± 1%   150µs ± 1%   -2.73%  (p=0.000 n=10+10)                                                                                                                                             
micro/memory_grow_mstore/nogrow   493µs ± 0%   469µs ± 0%   -4.88%  (p=0.000 n=10+9)                                                                                                                                              
micro/memory_grow_mstore/by1      502µs ± 1%   482µs ± 1%   -4.01%  (p=0.000 n=10+8)                                                                                                                                              
micro/memory_grow_mstore/by16     593µs ± 0%   581µs ± 1%   -2.03%  (p=0.000 n=10+10)                                                                                                                                             
micro/memory_grow_mstore/by32     642µs ± 1%   648µs ± 0%   +1.08%  (p=0.000 n=10+10)                                                                                                                                             
micro/JUMPDEST_n0/empty          3.45ms ± 0%  3.14ms ± 0%   -8.96%  (p=0.000 n=8+10)
main/sha1_divs/empty              227µs ± 1%   213µs ± 0%   -6.54%  (p=0.000 n=10+9)
main/sha1_divs/1351              4.60ms ± 0%  4.29ms ± 0%   -6.76%  (p=0.000 n=10+9)
main/sha1_divs/2737              8.94ms ± 0%  8.35ms ± 0%   -6.69%  (p=0.000 n=10+9)
main/sha1_divs/5311              17.5ms ± 1%  16.3ms ± 0%   -6.62%  (p=0.000 n=10+10)
main/weierstrudel/0               491µs ± 0%   478µs ± 0%   -2.77%  (p=0.000 n=9+10)
main/weierstrudel/1              1.05ms ± 1%  0.99ms ± 0%   -5.38%  (p=0.000 n=10+10)
main/weierstrudel/3              1.65ms ± 0%  1.56ms ± 0%   -5.21%  (p=0.000 n=9+10)
main/weierstrudel/9              3.42ms ± 0%  3.25ms ± 0%   -5.04%  (p=0.000 n=10+9)
main/weierstrudel/14             4.89ms ± 1%  4.64ms ± 0%   -5.01%  (p=0.000 n=10+10)
main/sha1_shifts/empty            180µs ± 0%   161µs ± 0%  -10.62%  (p=0.000 n=9+10)
main/sha1_shifts/1351            3.68ms ± 1%  3.28ms ± 1%  -10.75%  (p=0.000 n=9+10)
main/sha1_shifts/2737            7.13ms ± 0%  6.40ms ± 1%  -10.21%  (p=0.000 n=9+10)
main/sha1_shifts/5311            13.9ms ± 0%  12.4ms ± 0%  -10.78%  (p=0.000 n=10+9)
main/blake2b_huff/empty           110µs ± 0%   105µs ± 0%   -4.81%  (p=0.000 n=10+10)
main/blake2b_huff/2805nulls      2.10ms ± 1%  1.99ms ± 1%   -5.50%  (p=0.000 n=10+10)
main/blake2b_huff/5610nulls      4.08ms ± 1%  3.86ms ± 1%   -5.42%  (p=0.000 n=10+10)
main/blake2b_huff/8415nulls      5.96ms ± 1%  5.64ms ± 0%   -5.41%  (p=0.000 n=10+10)
main/blake2b_shifts/2805nulls    16.7ms ± 0%  15.5ms ± 1%   -7.27%  (p=0.000 n=8+10)
main/blake2b_shifts/5610nulls    33.3ms ± 0%  30.8ms ± 0%   -7.36%  (p=0.000 n=8+8)
main/blake2b_shifts/8415nulls    49.9ms ± 2%  45.9ms ± 1%   -7.94%  (p=0.000 n=10+10)
[Geo mean]                       1.58ms       1.49ms        -5.60%
```